took all attributes from various card types and consolidated them into Card.java
