+++ src/main/resources/templates/index.html
@@ -10,7 +10,7 @@
 
     <nav class="navbar navbar-expand-lg navbar-light bg-light">
         <div class="container-fluid">
-            <a class="navbar-brand" href="/templates/index.html"><img src="/static/logos/android-netrunner-logo.png" width="300"></a>
+            <a class="navbar-brand" href="#"><img src="/resources/static/logos/android-netrunner-logo.png" width="300"></a>
             <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
                 <span class="navbar-toggler-icon"></span>
             </button>
